Measurement for reaching its highest altitude.
firework:
since i turn objects null when they reach their highest altitudes I
check that before I add them to the scene the object still exists.

particles:
changed to 2D so that it can run better since there are a lot of not
destroyed objects populating the scene.
Check that the object is not null when adding it to the scene.