Added random colors to the particles.
Fireworks now have random hex colors.
In index.html:
I changed the canvas so that it preserves the drawing buffer. This way
we can have plane on top of the canvas rendering with a different
opacity value which makes it look as if we have a trail that fades away.